[#3] move. AddStoreRequest 기존 application -> adapter 이동
